Format change
Loading/Saving novels will now be saved into an ADL file with the extension .ADL.

An ADL file is just all the files and directories zipped together.

The ADL file also contains a command, which, in the future, will be used to allow people to drag n' drop files onto the executable to automatically update the file.

ADL also contains the url and date that the novel was last updated/created.

GetText has been converged in to their respective downloader classes.