Simplification and TwistMoe Search support work
Argument Object is now just a "holder" for argumentList. Streamlines argument settings, so I no longer have to add variables to the parsing function and ArgumentObject.

Twist.Moe Search support is getting closer to being done, and all that is left is to create a similarity algorithm to compare given title to all anime. (Plan to generate a list of "similar anime" from the first character that the anime starts with. E.x if search term starts