Standardization beginning.
First step to standardizing functions and organization between anime and novel options of the downloader.

Expect the novel downloader to be entirely rewritten, following the model of the anime downloader. All novel sites will be self contained, and the "Book" class will no longer be the one calling the shots.