Initial transfer of video from HentaiVideo and the bloated JsonMain file.
The plan is to to implement a similar system as what exists in novels with a general MetaData object. There will be a Video class similar to that of the Book class, which will store information specific to video, e.x manifest, links. This should provice a much cleaner solution.